(PC-7905): Move Provider to models file
