+++ src/pcapi/local_providers/provider_api/provider_api_stocks.py
@@ -2,7 +2,7 @@ import logging
 
 from sqlalchemy.orm.util import aliased
 
-from pcapi.core.providers.provider import Provider
+from pcapi.core.providers.models import Provider
 from pcapi.core.providers.venue_provider import VenueProvider
 from pcapi.models import db
