+++ src/pcapi/repository/repository.py
@@ -3,9 +3,9 @@ from sqlalchemy.exc import IntegrityError
 from sqlalchemy.exc import InternalError
 
 from pcapi.models import ApiErrors
-from pcapi.models import PcObject
 from pcapi.models.db import Model
 from pcapi.models.db import db
+from pcapi.models.pc_object import PcObject
 from pcapi.validation.models import entity_validator
