+++ tests/models/pc_object_test.py
@@ -13,12 +13,12 @@ from sqlalchemy.dialects.postgresql import UUID
 from pcapi.core.users.models import User
 from pcapi.model_creators.generic_creators import create_stock
 from pcapi.models import Offer
-from pcapi.models import PcObject
 from pcapi.models import ThingType
 from pcapi.models.api_errors import DateTimeCastError
 from pcapi.models.api_errors import DecimalCastError
 from pcapi.models.api_errors import UuidCastError
 from pcapi.models.db import Model
+from pcapi.models.pc_object import PcObject
 from pcapi.routes.serialization import serialize
 from pcapi.utils.human_ids import NonDehumanizableId
 from pcapi.utils.human_ids import dehumanize
